Moved tests from common/ into lms/
